requirements.txt
aniso8601==10.0.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
Flask==3.1.3
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-WTF==1.2.2
greenlet==3.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pip==23.2.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
setuptools==82.0.1
six==1.17.0
SQLAlchemy==2.0.49
sqlalchemy-serializer==1.6.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
WTForms==3.2.1